requirements.txt
webrtcvad
pydub
onnxruntime
transformers
torch
torchaudio
soundfile
python-dotenv